Refactor monolithic main.py into modular package structure
BREAKING CHANGE: Code organization completely restructured

## 📦 New Package Structure:
- **cve_tracker/**: Core package with logical separation
  - `package_mapping.py`: Maps packages to repository names (54 lines)
  - `security_scoring.py`: Security relevance scoring algorithms (116 lines)
  - `github_search.py`: GitHub API search and discovery (248 lines)
  - `__init__.py`: Clean package exports

## 📉 Reduced Complexity:
- **main.py**: 553 → 151 lines (73% reduction)
- **Separated concerns**: Each module has single responsibility
- **Centralized security keywords**: SECURITY_KEYWORDS constant
- **Clean imports**: All functions available via `from cve_tracker import ...`

## ✅ Benefits:
- **Maintainability**: Easier to understand and modify individual components
- **Testability**: Modules can be tested in isolation
- **Reusability**: Components can be imported independently
- **Scalability**: Easy to add new features without bloating main.py

## 🔧 Updated:
- Tests now import from `cve_tracker` modules
- Documentation reflects new structure
- All 27 tests still pass
- Code quality checks pass (ruff, mypy, pytest)

No functional changes - pure refactoring for better code organization.
